refactor: 미사용 이미지 주기적 삭제 기능 수정

+++ be/src/main/java/kr/codesquad/jazzmeet/image/service/ImageScheduler.java
@@ -18,8 +18,9 @@ public class ImageScheduler {
 	public void deleteNotRegisteredImages() {
 		List<String> imageUrls = imageService.findNotRegisteredImageUrls();
 
-		List<String> deletedImageUrls = cloudService.deleteImages(imageUrls);
-
-		imageService.deleteImagesByUrls(deletedImageUrls);
+		if (!imageUrls.isEmpty()) {
+			List<String> deletedImageUrls = cloudService.deleteImages(imageUrls);
+			imageService.deleteImagesByUrls(deletedImageUrls);
+		}
 	}
 }
